Added cast to suppress compiler warnings

+++ NSDate+Helper.m
@@ -80,7 +80,7 @@ static NSDateFormatter *displayFormatter;
 			text = @"Yesterday";
 			break;
 		default:
-			text = [NSString stringWithFormat:@"%d days ago", daysAgo];
+			text = [NSString stringWithFormat:@"%lu days ago", (unsigned long)daysAgo];
 	}
 	return text;
 }
